Update patch set 3
Patch Set 3: Code-Review+1

Good to go for next nightly. Only missing All Apps button, still. Needs mention on next changelog how to solve it:

SlimLauncher currently has an issue with the "All Apps" button on Tablets, it is missing by default. To add the "All Apps" button longpress on your Homescreen and click on "WIDGETS". Scroll down to "SlimLauncher", touch and hold it to place it on your Homescreen. Now choose "All apps" from the SlimLauncher shortcuts menu. Now you longpress on the new generated Shortcut and place it in the Launcher Dock to access it from every Homescreen.

Patch-set: 3
Label: Code-Review=+1